app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.9
django-debug-toolbar==4.4.6
django-rosetta==0.10.0
idna==3.10
polib==1.2.0
psycopg==3.2.2
psycopg2-binary==2.9.9
pymemcache==3.5.2
python-dotenv==1.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
gunicorn==21.2.0